+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [## [unreleased]]
 ### Added
-- additional manufacturer information as attributes (@Pfabeck)
+- additional manufacturer information as attributes (@Pfabeck & @ugurkankya)
 
 ## [## [5.3.2]]
 ### Fixed