+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [## [unreleased]]
+## [## [5.3.3]]
 ### Added
 - additional manufacturer information as attributes (@Pfabeck & @ugurkankya)
-- set plentymarkets system currency for payments
+- set plentymarkets system currency for payments @Pfabeck
 
 ## [## [5.3.2]]
 ### Fixed